BOAC-4036, tame the StudentProfileHeader component and tidy up formatting
